requirements.txt
# ===========================
# Data Collection
# ===========================
vnstock3>=0.3
beautifulsoup4>=4.12
requests>=2.31
lxml>=5.1

# ===========================
# Data Processing
# ===========================
pandas>=2.1
numpy>=1.24
ta>=0.11
scikit-learn>=1.3

# ===========================
# Deep Learning
# ===========================
torch>=2.1

# ===========================
# NLP / Sentiment Analysis
# ===========================
transformers>=4.36
underthesea>=6.8

# ===========================
# Web Dashboard
# ===========================
streamlit>=1.31
plotly>=5.18

# ===========================
# Utilities
# ===========================
loguru>=0.7
python-dotenv>=1.0
